refactoring
move extensions
remove useless files
remove multidex
