+++ app/src/main/java/com/semisonfire/cloudgallery/ui/photo/PhotoDetailPresenterImpl.kt
@@ -3,6 +3,7 @@ package com.semisonfire.cloudgallery.ui.photo
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import com.semisonfire.cloudgallery.core.data.model.Photo
+import com.semisonfire.cloudgallery.core.logger.printThrowable
 import com.semisonfire.cloudgallery.core.mvp.MvpPresenter
 import com.semisonfire.cloudgallery.core.presentation.BasePresenter
 import com.semisonfire.cloudgallery.ui.disk.data.DiskRepository
@@ -11,7 +12,6 @@ import com.semisonfire.cloudgallery.ui.trash.data.TrashRepository
 import com.semisonfire.cloudgallery.utils.FileUtils
 import com.semisonfire.cloudgallery.utils.background
 import com.semisonfire.cloudgallery.utils.foreground
-import com.semisonfire.cloudgallery.utils.printThrowable
 import io.reactivex.Observable
 import io.reactivex.Single
 import java.net.URL
