+++ app/src/main/java/com/semisonfire/cloudgallery/ui/trash/TrashPresenterImpl.kt
@@ -1,6 +1,7 @@
 package com.semisonfire.cloudgallery.ui.trash
 
 import com.semisonfire.cloudgallery.core.data.model.Photo
+import com.semisonfire.cloudgallery.core.logger.printThrowable
 import com.semisonfire.cloudgallery.core.mvp.MvpPresenter
 import com.semisonfire.cloudgallery.core.presentation.BasePresenter
 import com.semisonfire.cloudgallery.ui.disk.LIMIT
@@ -8,7 +9,6 @@ import com.semisonfire.cloudgallery.ui.trash.data.TrashRepository
 import com.semisonfire.cloudgallery.ui.trash.model.TrashViewModel
 import com.semisonfire.cloudgallery.utils.background
 import com.semisonfire.cloudgallery.utils.foreground
-import com.semisonfire.cloudgallery.utils.printThrowable
 import io.reactivex.Observable
 
 interface TrashPresenter : MvpPresenter<TrashViewModel, TrashView> {
