+++ app/src/main/java/com/semisonfire/cloudgallery/utils/AndroidUtils.kt
@@ -1,8 +1,14 @@
 package com.semisonfire.cloudgallery.utils
 
 import android.content.Context
+import android.graphics.PorterDuff
+import android.graphics.drawable.ColorDrawable
+import android.support.annotation.ColorInt
+import android.support.annotation.ColorRes
 import android.support.annotation.DimenRes
 import android.support.annotation.StringRes
+import android.support.v4.content.ContextCompat
+import android.view.Menu
 import android.widget.Toast
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
@@ -17,6 +23,26 @@ fun Context.dip(value: Float): Int = dipF(value).toInt()
 fun Context.sp(value: Int): Int = (value * resources.displayMetrics.scaledDensity).toInt()
 fun Context.sp(value: Float): Int = (value * resources.displayMetrics.scaledDensity).toInt()
 fun Context.dimen(@DimenRes resource: Int): Int = resources.getDimensionPixelSize(resource)
+
+fun Context.string(@StringRes resId: Int): String = getString(resId)
+
+fun Context.color(@ColorRes colorRes: Int) = ContextCompat.getColor(this, colorRes)
+fun Context.colorDrawable(@ColorInt color: Int) = ColorDrawable(color)
+fun Context.colorResDrawable(@ColorRes colorRes: Int): ColorDrawable {
+    val color = color(colorRes)
+    return ColorDrawable(color)
+}
+
+/** Change menu items icon color  */
+fun setMenuIconsColor(menu: Menu, @ColorInt color: Int) {
+    for (i in 0 until menu.size()) {
+        val drawable = menu.getItem(i).icon
+        if (drawable != null) {
+            drawable.mutate()
+            drawable.setColorFilter(color, PorterDuff.Mode.SRC_IN)
+        }
+    }
+}
 
 fun Context.shortToast(text: String) =
     Toast.makeText(this, text, Toast.LENGTH_SHORT).show()
