--- app/src/main/java/com/semisonfire/cloudgallery/utils/StringUtils.kt
@@ -1,6 +1,0 @@
-package com.semisonfire.cloudgallery.utils
-
-import android.content.Context
-import android.support.annotation.StringRes
-
-fun Context.string(@StringRes resId: Int): String = getString(resId)
